requirements.txt
requests
pyyaml
websocket
websocket_client